// src/graphql/socialMedia/getSocialMedia.generated.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /**
   * The `Date` scalar type represents a Date
   * value as specified by
   * [iso8601](https://en.wikipedia.org/wiki/ISO_8601).
   */
  Date: { input: any; output: any; }
  /**
   * The `DateTime` scalar type represents a DateTime
   * value as specified by
   * [iso8601](https://en.wikipedia.org/wiki/ISO_8601).
   */
  DateTime: { input: any; output: any; }
  /**
   * The `GenericScalar` scalar type represents a generic
   * GraphQL scalar value that could be:
   * String, Boolean, Int, Float, List or Object.
   */
  GenericScalar: { input: any; output: any; }
  /**
   * Allows use of a JSON String for input / output from the GraphQL schema.
   *
   * Use of this type is *not recommended* as you lose the benefits of having a defined, static
   * schema (one of the key benefits of GraphQL).
   */
  JSONString: { input: any; output: any; }
  /**
   * The `Time` scalar type represents a Time value as
   * specified by
   * [iso8601](https://en.wikipedia.org/wiki/ISO_8601).
   */
  Time: { input: any; output: any; }
  /**
   * Create scalar that ignores normal serialization/deserialization, since
   * that will be handled by the multipart request spec
   */
  Upload: { input: any; output: any; }
};
// ...
export type BrandNodeConnection = {
  __typename?: 'BrandNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<BrandNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars['Int']['output']>;
};
// ...
export type CartNodeConnection = {
  __typename?: 'CartNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<CartNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars['Int']['output']>;
};
// ...
export type CompanyTypeNodeConnection = {
  __typename?: 'CompanyTypeNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<CompanyTypeNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
};
// ...
export type ContactNodeConnection = {
  __typename?: 'ContactNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<ContactNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
};
// ...
export type DocumentNodeConnection = {
  __typename?: 'DocumentNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<DocumentNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars['Int']['output']>;
};
// ...
export type DownloadLogNodeConnection = {
  __typename?: 'DownloadLogNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<DownloadLogNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars['Int']['output']>;
};
// ...
export type EmailSubscriberType = {
  __typename?: 'EmailSubscriberType';
  createdAt: Scalars['DateTime']['output'];
  email: Scalars['String']['output'];
  firstName: Scalars['String']['output'];
  id: Scalars['ID']['output'];
  instance: InstanceNode;
  lastName: Scalars['String']['output'];
};
// ...
export type FaqNodeConnection = {
  __typename?: 'FAQNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<FaqNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
};
// ...
export type FlatpageNodeConnection = {
  __typename?: 'FlatpageNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<FlatpageNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars['Int']['output']>;
};
// ...
export type GlobalDataNodeConnection = {
  __typename?: 'GlobalDataNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<GlobalDataNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
};
// ...
export type HomeSliderNodeConnection = {
  __typename?: 'HomeSliderNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<HomeSliderNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
};
// ...
export type InstanceNode = Node & {
  __typename?: 'InstanceNode';
  brands: BrandNodeConnection;
  cartSet: CartNodeConnection;
  cartitemSet: ItemNodeConnection;
  claimSet: GlobalDataNodeConnection;
  code?: Maybe<Scalars['String']['output']>;
  companytypes: CompanyTypeNodeConnection;
  contacts: ContactNodeConnection;
  documents: DocumentNodeConnection;
  downloadlogSet: DownloadLogNodeConnection;
  emailsubscriberSet: Array<EmailSubscriberType>;
  facebookUrl: Scalars['String']['output'];
  faqs: FaqNodeConnection;
  flatpageSet: FlatpageNodeConnection;
  homesliderSet: HomeSliderNodeConnection;
  /** The ID of the object. */
  id: Scalars['ID']['output'];
  instagramUrl: Scalars['String']['output'];
  instance: ServiceContactNodeConnection;
  jobApplications: JobApplicationNodeConnection;
  linkedinUrl: Scalars['String']['output'];
  logo?: Maybe<Scalars['String']['output']>;
  markets: MarketNodeConnection;
  name: Scalars['String']['output'];
  orderitemSet: OrderItemNodeConnection;
  orders: OrderNodeConnection;
  posts: PostNodeConnection;
  products: ProductNodeConnection;
  projects: ProjectNodeConnection;
  projectsInstance: PromotedProjectNodeConnection;
  /** Please insert the emails separated for commas used in samples request in form No! shopping cart */
  sampleEmails?: Maybe<Scalars['String']['output']>;
  samplescontact: SampleContactNodeConnection;
  services: ServiceNodeConnection;
  stores: StoreNodeConnection;
  storesInstance: OfficeNodeConnection;
  twitterUrl: Scalars['String']['output'];
  userProfiles: ProfileNodeConnection;
  users: UserNodeConnection;
  vacancies: VacantNodeConnection;
};
// ...
export type ItemNodeConnection = {
  __typename?: 'ItemNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<ItemNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars['Int']['output']>;
};
// ...
export type JobApplicationNodeConnection = {
  __typename?: 'JobApplicationNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<JobApplicationNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
};
// ...
export type MarketNodeConnection = {
  __typename?: 'MarketNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<MarketNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
};
// ...
export type Node = {
  /** The ID of the object. */
  id: Scalars['ID']['output'];
};
// ...
export type OfficeNodeConnection = {
  __typename?: 'OfficeNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<OfficeNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
};
// ...
export type OrderItemNodeConnection = {
  __typename?: 'OrderItemNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<OrderItemNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
};
// ...
export type OrderNodeConnection = {
  __typename?: 'OrderNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<OrderNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
};
// ...
export type PostNodeConnection = {
  __typename?: 'PostNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<PostNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars['Int']['output']>;
};
// ...
export type ProductNodeConnection = {
  __typename?: 'ProductNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<ProductNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars['Int']['output']>;
};
// ...
export type ProfileNodeConnection = {
  __typename?: 'ProfileNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<ProfileNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
};
// ...
export type ProjectNodeConnection = {
  __typename?: 'ProjectNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<ProjectNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars['Int']['output']>;
};
// ...
export type PromotedProjectNodeConnection = {
  __typename?: 'PromotedProjectNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<PromotedProjectNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars['Int']['output']>;
};
// ...
export type Query = {
// ...
  instance?: Maybe<InstanceNode>;
// ...
};
// ...
export type SampleContactNodeConnection = {
  __typename?: 'SampleContactNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<SampleContactNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
};
// ...
export type ServiceContactNodeConnection = {
  __typename?: 'ServiceContactNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<ServiceContactNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
};
// ...
export type ServiceNodeConnection = {
  __typename?: 'ServiceNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<ServiceNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars['Int']['output']>;
};
// ...
export type StoreNodeConnection = {
  __typename?: 'StoreNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<StoreNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars['Int']['output']>;
};
// ...
export type UserNodeConnection = {
  __typename?: 'UserNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<UserNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
};
// ...
export type VacantNodeConnection = {
  __typename?: 'VacantNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<VacantNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
};
// ...
export type GetSocialMediaQuery = { __typename?: 'Query', instance?: { __typename?: 'InstanceNode', id: string, facebookUrl: string, instagramUrl: string, linkedinUrl: string } | null };